Day 683
A desktop screenshot of my 683rd consecutive day on Khan Academy, in PNG format. I got upset, as I couldn't find it originally. I just assumed someone could recreate it, so I moved on. It would have disappointed me, but when I went on, I found that it was grouped with the 4.11.2019 folder, so luckily it wasn't lost.